moved WoT-related logic to its own .ts file
